+++ src/ampworks/__init__.py
@@ -1,19 +1,19 @@
 """
 Summary
 =======
-``ampworks`` is a collection of tools designed to visualize and process
+`ampworks` is a collection of tools designed to visualize and process
 experimental battery data. It provides routines for degradation mode analysis,
 parameter extraction from common protocols (e.g., GITT, ICI, etc.), and more.
 These routines provide key properties for life and physics-based models (e.g.,
 SPM and P2D). Graphical user interfaces (GUIs) are available for some of the
-analyses. See a list of the GUI-based applications by running ``ampworks -h``
+analyses. See a list of the GUI-based applications by running `ampworks -h`
 in your terminal after installation.
 
-Note: ``ampworks`` is in early development. The API may change as it matures.
+Note: `ampworks` is in early development. The API may change as it matures.
 
 Accessing the Documentation
 ---------------------------
-Documentation is accessible via Python's ``help()`` function which prints
+Documentation is accessible via Python's `help()` function which prints
 docstrings from a package, module, function, class, etc. You can also access
 the documentation by visiting the website, hosted on Read the Docs. The website
 includes search functionality and more detailed examples.
@@ -36,6 +36,7 @@ __all__ = [
     'read_csv',
     'read_excel',
     'read_table',
+    'ocv',
     'ici',
     'gitt',
     'dqdv',
@@ -50,12 +51,13 @@ __all__ = [
 
 if TYPE_CHECKING:  # pragma: no cover
     from ampworks import (
-        ici, gitt, dqdv, hppc, utils, datasets, mathutils, plotutils,
+        ocv, ici, gitt, dqdv, hppc, utils, datasets, mathutils, plotutils,
     )
 
 
 # Lazily load submodules/subpackages
 _lazy_modules = {
+    'ocv': 'ampworks.ocv',
     'ici': 'ampworks.ici',
     'gitt': 'ampworks.gitt',
     'dqdv': 'ampworks.dqdv',