+++ src/ampworks/datasets/__init__.py
@@ -1,5 +1,5 @@
 """
-This module provides access to example datasets bundled with ``ampworks``. The
+This module provides access to example datasets bundled with `ampworks`. The
 available functions allow users to list, download, or load in example datasets.
 The datasets are used in tutorials and tests. They provide a convenient intro
 to package functions without the overhead of requiring users to perform their
@@ -8,11 +8,11 @@ own experiments.
 Most datasets were created using physics-based models like the pseudo-2D model.
 A brief description of each dataset is given below:
 
-1. ``gitt_charge`` - example GITT data (using charge/rest sequences)
-2. ``gitt_discharge`` - example GITT data (using discharge/rest sequences)
-3. ``ici_charge`` - example ICI data (using charge/rest sequences)
-4. ``ici_discharge`` - example ICI data (using discharge/rest sequences)
-5. ``hppc_discharge`` - example HPPC data (using discharge sequences)
+1. `gitt_charge` - example GITT data (using charge/rest sequences)
+2. `gitt_discharge` - example GITT data (using discharge/rest sequences)
+3. `ici_charge` - example ICI data (using charge/rest sequences)
+4. `ici_discharge` - example ICI data (using discharge/rest sequences)
+5. `hppc_discharge` - example HPPC data (using discharge sequences)
 
 """
 
@@ -42,7 +42,7 @@ def list_datasets() -> list[str]:
     Returns
     -------
     names : list[str]
-        A list of example file names from an internal ``resources`` folder.
+        A list of example file names from an internal `resources` folder.
 
     """
     resources = pathlib.Path(os.path.dirname(__file__), 'resources')
@@ -56,7 +56,7 @@ def download_all(path: str | os.PathLike | None = None) -> None:
     Parameters
     ----------
     path : str or PathLike or None, optional
-        Path to parent directory where a new ``ampworks_datasets`` folder will
+        Path to parent directory where a new `ampworks_datasets` folder will
         be created and example datasets will be copied to. If None (default),
         the current working directory is used.
 
@@ -86,7 +86,7 @@ def load_datasets(*names: str) -> Dataset:
     -------
     datasets : Dataset or tuple[Dataset]
         A single dataset if one name, otherwise a tuple of datasets in the same
-        order as the given ``names``.
+        order as the given `names`.
 
     Raises
     ------