+++ src/ampworks/ici/_extract_params.py
@@ -41,7 +41,7 @@ def extract_params(data: Dataset, radius: float, tmin: float = 1,
     ----------
     data : Dataset
         The sliced ICI data to process. Must have, at a minimum, columns for
-        ``{'Seconds', 'Amps', 'Volts'}``. See notes for more information.
+        `{'Seconds', 'Amps', 'Volts'}`. See notes for more information.
     radius : float
         The representative particle radius of your active material (in meters).
         It's common to use D50 / 2, i.e., the median radius of a distribution.
@@ -61,7 +61,7 @@ def extract_params(data: Dataset, radius: float, tmin: float = 1,
         Table of parameters. Columns include 'SOC' (state of charge, -), 'Ds'
         (diffusivity, m2/s), and 'Eeq' (equilibrium potential, V).
     stats : pd.DataFrame
-        Only returned if ``return_all=True``. Provides additional stats about
+        Only returned if `return_all=True`. Provides additional stats about
         each rest, including errors from the sqrt(t) vs. voltage regressions.
 
     Raises
@@ -74,7 +74,7 @@ def extract_params(data: Dataset, radius: float, tmin: float = 1,
     Notes
     -----
     Rests within the dataset are expected to have a current exactly equal to
-    zero. You can use ``data.zero_below('Amps', threshold)`` to manually zero
+    zero. You can use `data.zero_below('Amps', threshold)` to manually zero
     out currents below some tolerance, if needed. This should be done prior to
     passing in the dataset to this function.
 
@@ -85,11 +85,11 @@ def extract_params(data: Dataset, radius: float, tmin: float = 1,
     and discharge directions you should slice your data into two datasets and
     call this routine twice.
 
-    The default ``tmin`` and ``tmax`` values assume that rests occur for at
+    The default `tmin` and `tmax` values assume that rests occur for at
     least 10 s. You should adjust these accordingly if you use a protocol with
     shorter rests. Also, if a rest has fewer than two data points between the
-    set relative ``tmin`` and ``tmax`` then the linear regression performed to
-    find the diffusivity and equilibrium potential will return ``NaN`` for both.
+    set relative `tmin` and `tmax` then the linear regression performed to
+    find the diffusivity and equilibrium potential will return `NaN` for both.
 
     References
     ----------